feat(build): add pyproject.toml for build configuration
This commit adds a new file, pyproject.toml, which includes the build configuration for the project. It specifies that the project requires "hatchling" and uses "hatchling.build" as the build backend. Additionally, it sets up project metadata such as name, description, authors, and requires Python version 3.11.0 or higher. The commit also includes a Pyright tool configuration with venvPath set to "/Users/spedon/anaconda3/envs/drone2023/" and venv set to "drone2023".