refactor(ml): refactor model initialization and reinitialization
Refactored the `_NcnnModel` and `_OrtModel` classes to improve code organization and readability. Introduced a `reinit_with` method in both classes to handle model reinitialization with new parameters. Added a `clear` method in both classes to clear the model resources. Modified the `Model` class to include a `clear` method that calls the corresponding method in the underlying model instance. Also added a `reinit` method in the `Model` class to simplify reinitialization by calling the appropriate methods in the underlying model instance.