feat(core): add z_queue to Scheduler class
This commit adds a new attribute `z_queue` to the `Scheduler` class in the `vgcore/modules.py` file. The `z_queue` is initialized in the constructor and used in the `is_around()` method. This change allows for more accurate checking of drone positions.

refactor(utils): modify DroneInfo class constructor

The commit modifies the constructor of the `DroneInfo` class in the `utils.py` file. A new parameter, `land`, is added to the constructor signature and assigned as an instance variable. This change improves the flexibility of creating instances of the `DroneInfo` class by allowing for setting a default value for landing status.